// src/app/admin/core/models/fin-work-detail.ts
export class FinWorkDetail {
  work_detail_seq: number;
  work_seq: number;             
  year: number;                
  month: number;             
  expected_sales: number;      // 예상매출
  expected_purchase: number; // 예상매입
  sales: number;                   // 매출
  purchase: number;             // 매입
}